A4A Migrations: address review — dedicated agency-site-tags resource
Move the site-tags data layer out of the agency resource into its own
agency-site-tags resource (AgencySiteTag type, updateAgencySiteTags mutator,
agencySiteTagsMutation builder). Widen the migrations/site-tags builders to
accept number | undefined so the modals drop the agencyId casts, and type the
commission incentive statuses as a const union.
