+++ packages/api-core/src/agency/index.ts
@@ -1,3 +1,2 @@
 export * from './fetchers';
-export * from './mutators';
 export * from './types';
