+++ packages/api-core/src/index.ts
@@ -6,6 +6,7 @@ export * from './admin-bar';
 export * from './agency';
 export * from './agency-team';
 export * from './agency-sites';
+export * from './agency-site-tags';
 export * from './agency-woopayments';
 export * from './agency-migrations';
 export * from './akismet-api-key';
